Create iterate_through_index.py
This was written for the HTB CTF - Cyber Apocalypse 2024: Hacker Royale

CHALLENGE NAME
Character
Security through Induced Boredom is a personal favourite approach of mine. Not as exciting as something like The Fray, but I love making it as tedious as possible to see my secrets, so you can only get one character at a time!